+++ idm/ob.go
@@ -1,12 +1,46 @@
 package idm
 
 import (
+	"encoding/json"
 	"io/ioutil"
+	"strings"
 
 	"github.com/secureBankingAcceleratorToolkit/securebanking-openbanking-uk-fidc-initialiszer/am"
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
 )
+
+// ManagedObjectExists - checks if a managed object exists, must supply the object name
+func ManagedObjectExists(objectName string) bool {
+	path := "/openidm/config/managed"
+	result := &OBManagedObjects{}
+	b := am.Client.Get(path, map[string]string{
+		"Accept":           "application/json",
+		"X-Requested-With": "ForgeRock Identity Cloud Postman Collection",
+	})
+
+	err := json.Unmarshal(b, result)
+	if err != nil {
+		panic(err)
+	}
+
+	for _, o := range result.Objects {
+		zap.S().Infow("checking", "object", o)
+		if strings.Contains(o.Name, objectName) {
+			zap.L().Debug("ManagedObject " + objectName + " found")
+			return true
+		}
+	}
+	return false
+}
+
+// OBManagedObjects model
+type OBManagedObjects struct {
+	ID      string `json:"_id"`
+	Objects []struct {
+		Name string `json:"name"`
+	} `json:"objects"`
+}
 
 // AddOBManagedObjects -
 func AddOBManagedObjects() {
